requirements.txt
ansible==11.0.0
ansible-core==2.18.0
cffi==1.17.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
fonttools==4.55.0
grpcio==1.68.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mergexp==0.3.0
mpmath==1.3.0
numpy==2.1.3
packaging==24.2
pillow==11.0.0
protobuf==5.29.0
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
resolvelib==1.0.1
six==1.16.0
sympy==1.13.3
